Add Debuggable and EagerLoads traits for enhanced query debugging and eager loading
